repeat for reminders, requires more testing

+++ commands/remind.js
@@ -7,7 +7,7 @@ exports.execute = async (client, message, args) => {
 
     if (args[0]==="list") {
 
-        let txt = Object.entries(client.db.get("REMINDERS", message.author.id)).filter(r=>r[0]!="num").map(r=>`\`${r[0]}\` | \`${client.formatDate(new Date(r[1].time))}\`  |  \`${r[1].reason}\``).join("\n") || "You have no reminders!"
+        let txt = Object.entries(client.db.get("REMINDERS", message.author.id)).filter(r=>r[0]!="num").map(r=>`\`${r[0]}\` | \`${r[1].repeating ? `Every ${ms(r[1].time-r[1].created)}` : client.formatDate(new Date(r[1].time))}\`  |  \`${r[1].reason}\``).join("\n") || "You have no reminders!";
 
         let embed = new Discord.MessageEmbed()
         .setTitle("Reminders")
@@ -110,6 +110,12 @@ exports.execute = async (client, message, args) => {
         }
     }
 
+    let repeat = false;
+    if (args[0]==="repeat") {
+        args.shift();
+        repeat = true;
+    }
+
     let time;
     if(!args[0]){
         return message.channel.send("You must supply a time!")
@@ -133,9 +139,14 @@ exports.execute = async (client, message, args) => {
 
     let num = client.db.get("REMINDERS", `${message.author.id}.num`);
 
-    let timeout = client.setTimeout(()=>{
-        client.db.delete("REMINDERS", `${message.author.id}.${num}`)
-        delete client.remindtimers[`${message.author.id}-${num}`]
+    function remind() {
+        if(client.db.has("REMINDERS", `${message.author.id}.${num}`) && client.db.get("REMINDERS", `${message.author.id}.${num}.repeating`)) {
+            let timeout = client.setTimeout(remind, time)
+            client.remindtimers[`${message.author.id}-${num}`] = timeout;
+        } else {
+            client.db.delete("REMINDERS", `${message.author.id}.${num}`)
+            delete client.remindtimers[`${message.author.id}-${num}`]
+        }
         let embed = new Discord.MessageEmbed()
         .setTitle("Reminder")
         .setDescription(reason)
@@ -143,12 +154,18 @@ exports.execute = async (client, message, args) => {
         .setTimestamp(message.createdTimestamp)
         .setFooter(`Set on `)
         message.channel.send(message.author.toString(), {embed})
-    }, time)
+    }
 
-    message.channel.send(`Set reminder on ${client.formatDate(new Date(Date.now()+time))} for "${reason}"`)
+    let timeout = client.setTimeout(remind, time)
+
+    if(repeat) {
+        message.channel.send(`Set reminder to repeat every ${ms(time)} for "${reason}" you can cancel this with \`remind delete <id>\``)
+    } else {
+        message.channel.send(`Set reminder on ${client.formatDate(new Date(Date.now()+time))} for "${reason}"`)
+    }
 
     client.remindtimers[`${message.author.id}-${num}`] = timeout;
-    client.db.set("REMINDERS", {created: message.createdTimestamp, channel: message.channel.id, reason: reason, user: message.author.id, time: Date.now()+time}, `${message.author.id}.${num}`)
+    client.db.set("REMINDERS", {created: Date.now(), channel: message.channel.id, reason: reason, user: message.author.id, time: Date.now()+time, repeating: repeat}, `${message.author.id}.${num}`)
     client.db.inc("REMINDERS", `${message.author.id}.num`)
 };
   
@@ -158,6 +175,6 @@ exports.data = {
     category: "fun",
     name: "remind",
     desc: "Reminds you about something after some time",
-    usage: "remind <<time> | delete | edit | list> <reminder | id> [time | note] [modified time | note]",
+    usage: "remind [repeat] <<time> | delete | edit | list> <reminder | id> [time | note] [modified time | note]",
     perm: 0
 };
